move libverMkdirNested to libdir..
