+++ src/libver.c
@@ -34,74 +34,3 @@ libverNormalizedVersion(const char *version) {
     }
     return version;
 }
-
-/*
- * Verify if following absolute path points to directory
- * Returns operation status
- */
-static mkdir_status_t
-checkIfDirExists(const char *absDirPath) {
-    struct stat st = {0};
-    if (!scope_stat(absDirPath, &st)) {
-        if (S_ISDIR(st.st_mode)) {
-            return MKDIR_STATUS_EXISTS;
-        }
-        // TODO: add permissions to check if the file can be created there
-        return MKDIR_STATUS_NOT_ABSOLUTE_DIR;
-    }
-    // stat fails
-    return MKDIR_STATUS_OTHER_ISSUE;
-}
-
-/*
- * Create a directory in following absolute path creating any intermediate directories as necessary
- * Returns operation status
- */
-mkdir_status_t
-libverMkdirNested(const char *absDirPath) {
-    int mkdirRes = -1;
-    /* Operate only on absolute path */
-    if (absDirPath == NULL || *absDirPath != '/') {
-        return MKDIR_STATUS_NOT_ABSOLUTE_DIR;
-    }
-
-    mkdir_status_t res = checkIfDirExists(absDirPath);
-
-    /* exit if path exists */
-    if (res != MKDIR_STATUS_OTHER_ISSUE) {
-        return res;
-    }
-
-    char* tempPath = scope_strdup(absDirPath);
-    if (tempPath == NULL) {
-        goto end;
-    }
-
-    /* traverse the full path */
-    for (char* p = tempPath + 1; *p; p++) {
-        if (*p == '/') {
-            /* Temporarily truncate */
-            *p = '\0';
-            scope_errno = 0;
-            mkdirRes = scope_mkdir(tempPath, 0755);
-            /* scope_mkdir fails with error other than directory exists */
-            if (mkdirRes && (scope_errno != EEXIST)) {
-                 goto end;
-            }
-            *p = '/';
-        }
-    }
-
-    /* last element */
-    scope_errno = 0;
-    mkdirRes = scope_mkdir(tempPath, 0755);
-    if (mkdirRes && (scope_errno != EEXIST)) {
-        goto end;
-    }
-
-    res = MKDIR_STATUS_CREATED;
-
-end:
-    scope_free(tempPath);
-    return res;
-}
